# contracts/openapi/openapi.json
{"components":{"schemas":{"DataStruct":{"additionalProperties":false,"properties":{"key":{"type":"string"}},"required":["key"],"type":"object"},"ErrorDetail":{"additionalProperties":false,"properties":{"location":{"description":"Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'","type":"string"},"message":{"description":"Error message text","type":"string"},"value":{"description":"The value at the given location"}},"type":"object"},"ErrorModel":{"additionalProperties":false,"properties":{"detail":{"description":"A human-readable explanation specific to this occurrence of the problem.","examples":["Property foo is required but is missing."],"type":"string"},"errors":{"description":"Optional list of individual error details","items":{"$ref":"#/components/schemas/ErrorDetail"},"type":["array","null"]},"instance":{"description":"A URI reference that identifies the specific occurrence of the problem.","examples":["https://example.com/error-log/abc123"],"format":"uri","type":"string"},"status":{"description":"HTTP status code","examples":[400],"format":"int64","type":"integer"},"title":{"description":"A short, human-readable summary of the problem type. This value should not change between occurrences of the error.","examples":["Bad Request"],"type":"string"},"type":{"default":"about:blank","description":"A URI reference to human-readable documentation for the error.","examples":["https://example.com/errors/example"],"format":"uri","type":"string"}},"type":"object"},"SetSystemMetadataInputBody":{"additionalProperties":false,"properties":{"value":{"additionalProperties":{},"type":"object"}},"required":["value"],"type":"object"},"SetSystemMetadataOutputBody":{"additionalProperties":false,"properties":{"data":{"$ref":"#/components/schemas/DataStruct"},"success":{"type":"boolean"}},"required":["success","data"],"type":"object"},"StatusResponse":{"additionalProperties":false,"properties":{"data":{"$ref":"#/components/schemas/StatusResponseDataStruct"},"success":{"type":"boolean"}},"required":["success","data"],"type":"object"},"StatusResponseDataStruct":{"additionalProperties":false,"properties":{"status":{"type":"string"}},"required":["status"],"type":"object"}}},"info":{"title":"vkit-tango API","version":"1.0.0"},"openapi":"3.1.0","paths":{"/api/v1/status":{"get":{"operationId":"v1_get_system_status","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get system status"}},"/api/v1/system-metadata/{key}":{"put":{"operationId":"v1_set_system_metadata","parameters":[{"in":"path","name":"key","required":true,"schema":{"minLength":1,"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetSystemMetadataInputBody"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetSystemMetadataOutputBody"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Set system metadata"}}}}